Change EventParser::Append() to accept const SocketDataEvent&
Summary:
One unfortunate discovery is that the string has to be removed a prefix to match the offset
semantic.

http_parse_test.cc changed to test Parse() instead of EventParser::ParseMessages(), as those tests
are not verifying the part of ParseMessages() behaviors that is outside of Parse().

Need this to introduce tid to the message.

Test Plan: Jenkins

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2522

GitOrigin-RevId: 0a68ecf934845d3b6394f27e134e79291b7db18b