Fix leader election resignation
Summary:
Not sure if this will completely solve the dangling references issue, but it is possible for the metadata service to get out of sync when it is restarting.
defer does not actually run when a pod is terminating, so the metadata pod was never resigning as leader.
theres a period of time where a pod is terminating and a new one starts up, but can't yet become leader. during that time it may miss some events/miss syncing.

to actually be robust, we should periodically perform syncs on the metadata service. we currently only do it when the service starts up.

there was also a conditional error

Test Plan: ran skaffold

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2527

GitOrigin-RevId: c40b90c073ba61660fb979ef350d122196470253