[Clean-up/Style guide] Avoid using statements in header files
Summary:
Triggered by a different diff, there was a discussion about using statements and our style guide, and then I saw these violations.

Initially thought it was going to be a few small changes, but as I pulled on the thread, it unravelled.

In any case, since I went through the effort, here is a diff that changes some of our code-base to adhere to our style guide better.

Test Plan: Existing tests.

Reviewers: zasgar, philkuz, #engineering

Reviewed By: philkuz, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2534

GitOrigin-RevId: 4e62545f667239003537f7bb5699c461ab52dc91