Support None objects
Summary: In Python, every function call returns a value, even if a return type is not specified in the function. IN the latter case, the python function returns a None value. This will be used in the Result/Print() calls.

Test Plan: added test to make sure that None objects have methods, but cause compiler errors that are predictable. This will signal the user that they can't use the object they are trying to reference.

Reviewers: nserrino, michelle, zasgar, #engineering

Reviewed By: nserrino, #engineering

Subscribers: yzhao

Differential Revision: https://phab.corp.pixielabs.ai/D2547

GitOrigin-RevId: c1814e86588de579bc61cb247cc961bef4c45c71