Filter pods in metadata service update
Summary:
when sending out service updates to the agents, we need to filter out the pods in the service update which do not belong to the agent the update is being sent to.
we correctly do that when sending updates to an agent that has just registered. for subsequent updates, this was not happening.
because we are only running one pod/service in our system, we didn't encounter this issue that frequently. we would only be seeing it when a pod was dying and a new pod would be started up in a different node.

Test Plan: updated unit tests, ran skaffold

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2552

GitOrigin-RevId: 1e065937643cfcfbdd1014074c90cb1b266a73a7