Fix memSrcNode bug setting eos multiple times
Summary:
We were seeing a bug where there would occasionally be two row batches, one after the other, with eos set. eos set should only be set once.
this bug was most prominent in the new Kelvin work, because the grpcsink would try to close the writer twice and result in a crash.

with perfect timing, it is possible for the memorySourceNode to produce two rowbatches with eos set.
1. execGraph::Execute calls HasBatchesRemaining, sees it is true, so calls MemSrcNode's GenerateImpl.
2. MemSrcNode increments current_batch_ and calls HasBatchesRemaining is false. It sets the eos to true.
3. New rowbatch is added to the table by Stirling
4. execGraph::Execute calls HasBatchesRemaining and sees it is true again, so calls MemSrcNode's GenerateImpl
5. MemSrcNode outputs another rowbatch with eos true.

Instead, as soon as we set eos to true, we should make sure HasBatchesRemaining always returns false.

Test Plan: unit test

Reviewers: zasgar, nserrino, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2575

GitOrigin-RevId: 5b35e5e882879e2e5fbb35ba9c886fc207cfac78