Add rule to expand input columns in a map with keep input columns set to true
Summary:
This rule is going to make the refactor to the new form of maps
df['foo'] = df['bar'] + 2
a lot easier, because it will provide a shortcut to say that df takes all of the
input columns from the previous definition of df, and then it is adding a new column foo to that.
If there was already a 'foo' in df, it will be overwritten by the latest definition.

Test Plan: added a test

Reviewers: philkuz, zasgar, michelle, #engineering

Reviewed By: philkuz, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2570

GitOrigin-RevId: 4a2267026be85cd892ac25865f2314ac3388ee9c