Add QueryFlags object in querybroker
Summary:
we need to be able to parse query flags that are specified in the query string.
this diff adds a util for parsing the query string into a QueryFlags object. the object can be used to retrieve the set config values, or retrieve the default otherwise.

Test Plan: unit test

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Subscribers: philkuz

Differential Revision: https://phab.corp.pixielabs.ai/D2588

GitOrigin-RevId: 655d2191c603514d940049f748e14beb71f9bff8