PL-1177 - Add backend tracking events
Summary:
Added event tracking for Org, User, Site creation. All event tracking is added in the API server to prevent events from creeping
everywhere in the code base. For now since the API service is closest to the user it *should* be the best place to track this. We can revist this assumption if we run into other cases.

Test Plan: Updated tests, but did not add explicit tests for the tracking code.

Reviewers: michelle, philkuz, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PL-1177

Differential Revision: https://phab.corp.pixielabs.ai/D2621

GitOrigin-RevId: 0a5a66da25216b3d7755387b8816c0648182c7c6