PL-1168 Track Kelvins in metadata service
Summary:
We need to track which agents are Kelvins in the metadata service, since everytime a k8s update is sent we must send it to all of the kelvin nodes.
this diff updates the metadata service to store all of the kelvin nodes in a list.
this list is cleaned up when kelvins are deleted.
also added a method to get the list of kelvin ids. this will later be used for sending updates.

Test Plan: update unit tests, add more tests

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2628

GitOrigin-RevId: 09e02ae2687fcc5bddf36b895c935c6bce24b932