Add a Script view to the UI.
Summary:
Add a new 'scripts' route that is the shell of a code editor with tabs.

Implement an editor that manages opened tabs and persists the state to localStorage.

This is a work in progress, the new route is not linked from anywhere in the UI.

Test Plan: Tested the UI manually. The page is still work in progress.

Reviewers: michelle, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2623

GitOrigin-RevId: 240d944824c30a77587c6c331b573ccada137113