Update release notes
Summary:
We had two release note pages, so this diff removes one of them.
It was also rendering notes with \n incorrectly. updated the gatsby component to handle that correctly. followup diff will fix the shell script which generates the release to not generate \\n.

Test Plan: ran gatsby and checked

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2635

GitOrigin-RevId: 512ce10277567d4b0f784860dbefbddd9b7d6e07