Fix segment key issues caused by TSC optimization
Summary:
This should fix segment in the UI. The previous code was being optimized away by the compiler.
Actually, many variations of the previous code were optmized away, but since this is spread across modules it's safe.

Test Plan: Looked at the generated file manually and checked that the segment key is available for pre-processor replacement.

Reviewers: michelle, philkuz, malthus, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2649

GitOrigin-RevId: 2baf877885622671ded86dde1c3dea45d02f8c4a