Add aes-min submodule, with some tests
Summary:
With TLS, when we capture keys with uprobes, we actually get the expanded round keys.
For example, there are 11*16B such keys for AES-128.

But these keys are hard to use with existing crypto libraries, since their interface generally wants
the master key.

This diff includes a modified aes-min submodule. The aes-min has been modified to incorporate
a reverse key generator that goes from the expanded round keys back to the original key.

Next step is to use the recovered master key to show a proof-of-concept that we can decrypt captured traces.

Test Plan: A couple of AES-min tests added to show how the key recovery works.

Reviewers: zasgar, yzhao, #engineering

Reviewed By: zasgar, yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2656

GitOrigin-RevId: 48c809a1e7b85f7597f4948f034af5320960961c