Support PL module in the ast visitor
Summary: Use the PL module object and make it accessible for some stuff in the ast visitor. Don't replace everything yet to limit the surface area of the change, but this is the first step in any of this.

Test Plan: add expression test to make sure that everything works as expected

Reviewers: nserrino, michelle, zasgar, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2669

GitOrigin-RevId: 95d08258716e050042ce7118e93c1cff7b0a7940