Send all metadata state to Kelvin upon register
Summary:
When an agent registers with the metadata service, it sends the current metadata state that is relevant to that agent's node.
When Kelvin registers with the metadata service, we want to send it all metadata state regardless of Kelvin's node.
This diff adds that in.

We still need to add changes so that future k8s updates are sent to Kelvin (currently no updates are sent at all)

Test Plan: unit tests

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2684

GitOrigin-RevId: bb16c48b315be0ff55581462085dc94c06b1d83f