[PL-1085] Merge range into MemorySource
Summary: Remove range because it's not part of the pandas API. It already is required to be a child of MemorySource, so this solidifies it. I still have some cleanup todo, this is just to mark that I've pushed it up.

Test Plan: tests pass, should work. Need to rename rules and everything

Reviewers: nserrino, michelle, zasgar, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2713

GitOrigin-RevId: 1c5349cc1497ec3fe3efccf4e601b8acf5882dbf