Generalize rule to evaluate compile time expressions for maps, filters
Summary:
Previously the rule to evaluate compile time expressions only existed
for range. This diff generalizes it a bit more to make it do the same logic for
map and filter expressions. It will evaluate nested compile time expressions so
subargs will be evaluated even if the parent func is not a compile time func.
This is in support of the new range_agg syntax. Depends on D2715.

Test Plan: added tests

Reviewers: philkuz, #engineering

Reviewed By: philkuz, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2717

GitOrigin-RevId: 2f796f0a1666cfc92d69a9bedf6792ec219cf328