[PL-1169] Add test to make sure that default arguments can evaluate properly
Summary: I wanted to add a test to make sure that all of the default arguments will compile down to something. This doesn't guarantee that they will evaluate to what is expected, but at least they won't come up with some dumb errors from the default argument not having the correct syntax or something like that.

Test Plan: added tests for all existing objects and also have a test to make sure that the tests will fail when we expect them to.

Reviewers: nserrino, zasgar, michelle, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2721

GitOrigin-RevId: c4b5a02c70b1bbb8960702c125bf9ef58c3a2775