Fix the Drop() query Analzyer rule in Compiler
Summary:
the Drop() handling in the analyzer was wrong because we had a catchall for resolving Relations for any operator, and the drop to Map rule occured afterwards - meaning the information for the modified relation wouldn't actually appear until later and as a result cause
an error in Carnot.

Test Plan: added new tests to catch where this used to fail, they pass now.

Reviewers: nserrino, zasgar, michelle, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2729

GitOrigin-RevId: b49da4975504b532ed0511939df115a4791927b3