Update blocking agg benchmark
Summary:
To use QL.next. It still fails in a later test because the table size grows too large, beyond
table_store_table_size_limit. Let me know if people have feedback on the best way to address that issue.
I can see a few ways 1) change the default 2) change the test to not go to as high of a table
3) people can manually change table_store_table_size_limit to -1 when running the benchmark. But there
are probably options I am not thinking of.

Test Plan: ran it

Reviewers: philkuz, michelle, zasgar, #engineering

Reviewed By: philkuz, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2730

GitOrigin-RevId: e2f7ac29430384c86f965c1af280f86c63ef8627