PL-1038 Add display syntax to the system
Summary: TSI+ Haven't deprecated result() yet. Display syntax is not natively supported in pandas so we have to add it as a separate function that must be imported directly into the current Python environment if you were writing this in a python interpretor.

Test Plan: add test to make sure that display yields a memory source.

Reviewers: nserrino, zasgar, michelle, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PL-1038

Differential Revision: https://phab.corp.pixielabs.ai/D2736

GitOrigin-RevId: 80bdac8d5426b691e5c8c1e5d64da9ddb1c33a5d