SymbolAddress function to locate object addresses
Summary:
The golang interface type is actually an address in the symbol table.
This allows us to find the address for a type of interest.

Next step will be to inject this into the BPF code where interface types are used.

Test Plan: Test case included, that compares the result to nm's result.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2781

GitOrigin-RevId: 42f8e9e44c8797ecd84fa140528be3c8bc1a0652