Update Dataframe handler API to prepare new ASTvisitor architecture.
Summary: Old API had a pointer to dataframe and started causing problems where the referenced Dataframe was deleted before we use the function that the Dataframe created. Instead started pointing to the data that was within DataFrame in the bound functions, fixing the problem.

Test Plan: Update tests and they all work. This also fixes issues that came up later.

Reviewers: nserrino, zasgar, michelle, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2789

GitOrigin-RevId: c5b7e008bedbe129167117f9a1626fd856da90db