Integrate Dataframe subscript method as only option to handle subscripts in the code
Summary: This mostly removes special cases, the only exception is that we have to check to make sure that we can evaluate the subscript column if it's in a map or not.

Test Plan: Added new tests to make sure that maps that attempt to copy columns fail as well as metadata, which worked before.

Reviewers: nserrino, zasgar, michelle, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2794

GitOrigin-RevId: 80b128dccbdf21a304dcbfb7b0f83328218c6ca1