Fixed some bugs in union where empty row batches were not handled properly
Summary:
Union should be robust to 0-row RowBatches (both eos, not eos, a stream entirely
composed of 0-row RowBatches, and a stream with 0-row and >0 row RBs mixed in together.
This is especially important because GRPCSource will emit a 0-row RowBatch when it times out
on receiving data.

Test Plan: added tests

Reviewers: zasgar, michelle, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PL-583

Differential Revision: https://phab.corp.pixielabs.ai/D2790

GitOrigin-RevId: d9d28527aeea8d592a5cc735b4c12917fcd48750