Pass Distributed query flag to LogicalPlanner
Summary:
We need to pass a flag to the logical planner to tell it whether to use the distributed/non-distributed planner.
the query flag parsing was added in a past diff. this diff just passes that query flag to the logical planner (through a bunch of CGO stuff) to create the right planner.

Test Plan: existing tests should pass, ran it in skaffold and checked that #pl:set distributed_query=true tries to use the distributed planner.

Reviewers: philkuz, zasgar, #engineering

Reviewed By: philkuz, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2744

GitOrigin-RevId: 451142b5093d0afd0d805e0fef4b312c4aa03981