Add pixie cloud version to the UI.
Summary:
Use webpack to pass environment variables. Add the version number in the admin view if it is the prod version, or always show it on the bottom right if it is staging or local.

This depends on the the process calling webpack to pass the environment variables.

Test Plan: Tested manually

Reviewers: zasgar, michelle, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PL-1219

Differential Revision: https://phab.corp.pixielabs.ai/D2813

GitOrigin-RevId: c41fbc03173a1f8cfd9ea7bcee68a61d1e896119