Add SimulatedTimeSystem
Summary:
The simulatedTimeSystem is used for testing parts of the code that use a time system.
The SimulatedTimeSystem is initialized with the monotonic/system time that it should begin with and  has functions for manually updating the monotonic/system time.
there is also a sleep function for simulating sleep.
When time is advanced using any of the above methods, the time system manually goes through the active timers in the system and activates the ones that should have been triggered.

Test Plan: added test, check that is runs in asan and tsan

Reviewers: philkuz, zasgar, #engineering

Reviewed By: philkuz, zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2801

GitOrigin-RevId: be6f700f6de98dd067421b92b37f60a637c3e3f7