Move elf tools into own library inside stirling
Summary:
Starting to move experimental code into Stirling.

Elf tools seems like a good candidate to be its own library. We can add dwarf tools into the same directory in the future, so if there's a suggestion for a better common name, I'm open to it.

Test Plan: Moved tests as well.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2829

GitOrigin-RevId: e8956cd2a59494809ad1c6dff876c364d2023a59