PL-1224 Fix metadata dangling reference error
Summary:
we were running into dangling reference errors because we were not correctly reporting some objects as deleted.
i was making the incorrect assumption that the DeletionTimestampNS would be set for every deleted object... this actually only happens when the object is gracefully deleted.
instead, we should look at the event type received from the K8s watcher. if it is deleted and no deletiontimestamp is already set, we should just set it to Now().
Now, the metadata service should not be sending dead services with invalid references to any agents.

Test Plan: ran skaffold, made sure the endpoint was actually deleted. updated tests

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2881

GitOrigin-RevId: 5d3e9f7ef1606cb690367997b6c01e86cf4e2aa8