Move probe for operateHeaders() into go_grpc.c
Summary:
I planned to use the conn_fd2(), but realized later in the drafting that it requires user-space code
to setup the map.

Decide to keep it as a placeholder so that I could avoid reverting changes needed fro conn_fd2().

Once the interface is ready, remove conn_fd().

Test Plan:
Jenkins
Right now, we only verify that the prober returns header values from test log.

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2870

GitOrigin-RevId: ae73a9a8fad2de3b47c77e9cb8b6dbe265c7a1a6