Fix the vizier agents not updating.
Summary: It turns out the apollo useQuery only ever completes once... So it was not suitable to use in polling. Using the data returned by the useQuery hook fixes it.

Test Plan: Tried fixing the test but for it turns out the useQuery hooks requires a different MockedProvider, and updating that opened a new can of worms.

Reviewers: michelle, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PL-1199, PL-1227

Differential Revision: https://phab.corp.pixielabs.ai/D2885

GitOrigin-RevId: fa13169242a7daf7596e4c18bccb55324f3d5260