Address bug in GRPC Source
Summary:
There was logic to timeout and create a RowBatch after waiting a
certain amount of time in GRPCSource. However, due to a stale if clause above that logic,
it wasn't be executed under certain circumstances. It is removed now. The bug with the
empty mysql table is addressed by this but it will be good if we also address PL-1250.

Test Plan: ran the mysql query in the UI

Reviewers: zasgar, michelle, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2911

GitOrigin-RevId: e80ef4c2ff1da37d7e3cd0c42c2a9135bf963b28