PL-1226 part 1: Refactor IR::Clone() to be more generalizable
Summary:
Our current use case for IR node cloning was to clone an entire IR into a new graph.
However, as a result of PL-1226 we have the need to perform a copy of IR nodes into the graph
that they are already in. This refactor helps our approach to cloning cover both cases with
the same "node copying" logic.

more detail:
Previously, IR::Clone() was structured so that it calls out to each IRNode's implementation
of DeepCloneIntoImpl (via DeepCloneInto). This refactor changes the logic so that there is a top-level
IR::CopyNode, which calls out to each IRNode's implementation of CopyFromNode. The top-level
function decides whether or not to keep the same ID for the copies as the originals.
If they are part of the same graph, a new ID is used, otherwise the same ID is used.
By passing in the node we want to copy as the input, we are able to control the new ID vs old ID logic
entirely at the top level (avoiding boilerplate within every CopyFromNode).

Test Plan: existing tests of IR::Clone()

Reviewers: philkuz, zasgar, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PL-1226

Differential Revision: https://phab.corp.pixielabs.ai/D2917

GitOrigin-RevId: b9541fa9c2bfd63e40d27a1eded951d3c0202040