PL-1226 part 2: Add UDF type to registry, fix bug in RegistryInfo
Summary: This provides a method on RegistryInfo so that it's easy to look up the UDF type by name. This will allow pl.module to easily pull the appropriate UDF/UDAInfo in the upcoming changes to its parsing of udfs. Also this change exposed a big in the use of data() on a string view so that is fixed too.

Test Plan: existing

Reviewers: philkuz, zasgar, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PL-1226

Differential Revision: https://phab.corp.pixielabs.ai/D2926

GitOrigin-RevId: c594dd718db517d8f44fad034b1e1e6f2950f4cc