PL-1226 part 4: Support cloning nodes/dags which contain nodes with multiple parents
Summary:
In order to support multiple expressions used by the same parent, we need to support
cloning of DAGS like that. This updates the logic in Clone() and CopyNode() to add that functionality,
and refactors CloneTests to support this. It's possible some of the code in CloneTests will make
its way to IRNodes in the future to support doing comparisons of nodes.
Depends on D2917.

Test Plan: existing, added

Reviewers: philkuz, zasgar, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PL-1226

Differential Revision: https://phab.corp.pixielabs.ai/D2942

GitOrigin-RevId: 6229a15a44b73225b1e7e06a8f1400b389c27541