PL-1264 part 1: Support reused args in a Func
Summary:
TSIA. we should support cases like the following
a = 10
df['foo'] = a * a
This diff enables that support by cloning duplicate expressions in the same
function.

However in this case:
a = 10
df['foo'] = a #map_expr1
df['bar'] = a #map_expr2
a clone will not be performed, because map_expr1 and map_expr2 can both point to the node a without
duplicate edges.

Test Plan: added a test

Reviewers: philkuz, zasgar, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PL-1264

Differential Revision: https://phab.corp.pixielabs.ai/D2947

GitOrigin-RevId: a8f05b8fa5e5516fe866a4797cde0dfbe3a9a802