HTTP2 Tracer: Probe server-side headers
Summary:
Our operateHeaders probe was on the grpc http2Client function of that name.
This adds a probe on the grpc http2Server function as well.

The function signatures are mostly the same--at least the parts we care about--but the offset of the conn differs
in the respective structs.

So factor out some common code. The outer layer just extracts the relevant parameters, and then calls the inner code that does most of the work.

Test Plan: stirling_wrapper

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2951

GitOrigin-RevId: 591dd6e24e92c3c8316cde56b7a8b61459b1e033