PL-1226 part 3: Move Expression::ContainingOperator -> Column::ContainingOperators
Summary:
0) Remove limiting use of ContainingOperator() in the rule that resolves
datatypes and base UDA/UDF lookup on whether or not the registry has
that func listed as a UDA or UDF.
1) Now containing operator is only used for columns, so let's move
it there until others need it.
2) Consider the case
a = foo['a'] / 2
foo['b'] = a #op1
foo['c'] = a #op2
The column foo['a'] has two containing operators, op1 and op2.
However it has exactly one referenced operator, the operator which
produced foo as it was when a was defined. We now support columns having
multiple containing operators but enforce that there is only one
referenced operator.

Especially with item 0, we can now more easily support funcs that have
multiple containing operators as well.

If we think that all expressions should keep having a containing
operator lmk and I can put it back there.

Depends on D2926

Test Plan: existing

Reviewers: philkuz, zasgar, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PL-1226

Differential Revision: https://phab.corp.pixielabs.ai/D2929

GitOrigin-RevId: 61e4466bc82875e1e90a774aefb8b360ae82e83c