Fix CLI/Vizier build stamps
Summary:
Ishan found a bug where the version reported in the release CLI was "0.0.0-dev+Modified.63fcde0.20191217225735.0", which is the wrong build stamp.
in bazel/get_workspace_status, we use VERSION to be the build tag for the version string.
In https://phab.corp.pixielabs.ai/D2809, we mistakenly changed VERSION->SOURCE_VERSION to fix a bug in ci/build_cloud_artifacts.sh, which refers to a SOURCE_VERSION. Instead, we should make build_cloud_artifacts.sh refer to VERSION.

Test Plan: n/a

Reviewers: zasgar, oazizi, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2966

GitOrigin-RevId: 6e2dc5bf1e9ac49fa6f1c291d79676b7a8e24fdd