HTTP2 Tracer: Support stream IDs that go backwards, for robustness
Summary:
According to the HTTP2 spec, stream IDs should only go forwards.

But who knows what we will see in practice, due to data races and the fact that Stirling may only be getting a picture of what is happening mid-stream when we deploy after streams are already up.

So this change adds support for stream IDs that go backwards.

Test Plan: A couple of tests added.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2971

GitOrigin-RevId: 3c3e9b1afa5e9e41638cfe321d564c7a3132e5d9