Add tableName to query result
Summary:
now that we have the capability to show multiple tables in the UI, we also want to send the table names in the query result so that the UI can also show the table names.
modified queryresults proto to include table name.
updated graphql to include table name when resolving a query.
updated carnot so that it adds the table name to the query result in the CarnotQueryResult::ToProto method. since we actually intercept the sinks and update their name to "<query_id>_<num>" before executing the query, I needed to make a map from the new name to the old name.

Test Plan: added tests, updated tests

Reviewers: philkuz, #engineering, zasgar

Reviewed By: philkuz, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2974

GitOrigin-RevId: 2398ad771eb1fcfada31d1bcf0dfee42e350002f