Pass analyze/explain flags to plans
Summary:
we should be able to specify Analyze/Explain in the query params, and have these values passed down to the plans in the agent.
we had previously done this for just the "distributed" flag in a sort of hacky way.
this way cleans up how the distributed flag was set in the plan proto so that it makes sense for analyze/explain as well.

Test Plan: updated tests, added tests, ran in skaffold and logged out the logical plan to make sure the flags are set in the pb.

Reviewers: zasgar, philkuz, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2990

GitOrigin-RevId: 4755e6f3e1c9ce15b28942f6f7cb83168722a271