Fix Kelvin crash for failed pods
Summary:
I was running into an issue on my cluster where my Kelvin instance was always crashing at startup with the error:
```
F1223 19:09:08.837417     1 cgroup_metadata_reader.cc:99] Check failed: 0 Unknown QOS class
*** Check failure stack trace: ***
    @           0x9e873f  google::LogMessageFatal::~LogMessageFatal()
    @           0x6b4c6c  pl::md::CGroupMetadataReader::CGroupPodDirPath()
    @           0x6b5eb7  pl::md::CGroupMetadataReader::PodDirExists()
    @           0x6c09ad  pl::md::AgentMetadataStateManager::RemoveDeadPods()
    @           0x6c0159  pl::md::AgentMetadataStateManager::PerformMetadataStateUpdate()
```

Turns out that I had some pods with a "FAILED" status due to eviction. When a pod has "FAILED" status, it does not have a QOSClass.
This was resulting in a crash because `AgentMEtadataStateManager::RemoveDeadPods` would call `PodDirExists`, which would crash when QOSClass is unknown. QOSClass is only unknown when the pod status is failed.
Updated metadata service to send the pod status up to the agent. Now, we can check if the pod is failed before checking if PodDirExists.

However, it looks like we may be able to delete the RemoveDeadPods function. We should do some further examination to see if this is possible now. However, for the time-being, pod status seems like a reasonable thing to send to the agents anyway.

Test Plan: ran skaffold, made sure kelvin no longer crashes

Reviewers: zasgar, oazizi, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3009

GitOrigin-RevId: 2eb600869b4327755f72e5536a20e91deb2076b2