Add Cache struct to be used in metadata service
Summary:
etcd was reporting high CPU usage since we were making so many write requests per second.
as discussed, we should buffer these requests and instead make batch requests to etcd every 100ms or so.
to do so, we need a cache where which we can make reads and writes to. the cache is also responsible for flushing to etcd periodically.
Gets will get the key from the cache if it exists in the cache, otherwise it will reach out to etcd.
Sets will set the key/value in the cache.

Since we may switch out etcd in the future, we used a KeyValueStore interface so that the cache code is not reliant on etcd.
diff to implement the KeyValueStore interface for etcd is coming in a separate diff.

Test Plan: unit tests

Reviewers: zasgar, nserrino, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2992

GitOrigin-RevId: 1e32f03720e968f97241921303fb3ce3df2f030f