PL-1277 part 2: Add support for columns like df.foo
Summary:
Depends on D2986. We want to be able to support referring to columns like
df['foo'] as well as df.foo and we currently only have the former way. This diff adds that
support. Next diff will update preset queries and other end to end tests to use this new syntax
in some places.

Test Plan: added a test

Reviewers: philkuz, zasgar, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PL-1277

Differential Revision: https://phab.corp.pixielabs.ai/D2988

GitOrigin-RevId: cf402c4290976dbc3a623b1bd1d7aa4ce2b12054