PL-1277: Update some tests and preset queries to use df.foo syntax
Summary:
We now have extra unit test coverage for df.foo column syntax, in addition to the
explicit tests for that syntax that were added by previous diffs. carnot_test, which actually
executes queries, was also updated to use columns with that syntax. Finally, the preset queries
have been updated to use this syntax where possible (when special characters are not part of the
column name).

Please note: when this lands, people running the latest version of the UI will also need to run
a recent version of Vizier.

Test Plan: tsia

Reviewers: philkuz, michelle, zasgar, #engineering

Reviewed By: zasgar, #engineering

Subscribers: malthus, ishan

JIRA Issues: PL-1277

Differential Revision: https://phab.corp.pixielabs.ai/D3015

GitOrigin-RevId: 282e8e21971bd820f5f5a19aaee1be343839d38d