HTTP2 Tracer: Avoid memory consumption spikes when Stream ID jumps
Summary:
We currently use deques that are indexed by stream ID.
This means that if two stream IDs are far apart, the memory consumption will spike.

This diff limits the increase in deque size that is allowed by a stream ID jump.

This is mostly for safety, as the main reason this was notice was bad connection IDs that caused aliasing, but that has since been handled separately.

Test Plan: A couple of tests added for forward and backwards jumps.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3014

GitOrigin-RevId: 2b47cfbdbac3208ce2ab5d7b1083394386fe2428