PL-1296: Add utility to create zero-row row batches and check to make sure all columns have been allocated
Summary:
When someone does something like this
RowBatch rb(desc, 0);
Where desc has 2 columns
They may expect that the row batch will also have two columns.
However it does not have those columns until AddColumns is called.
To prevent future mistakes with this, I added a DCHECK in the RowBatch destructor to make sure the columns have all been added.

Also added a utility to create zero-row row batches of a certain schema, because that is something we need in a few places in the code
(those places did not call AddColumns so this is is also a bug fix).

Test Plan: added/existing

Reviewers: michelle, zasgar, philkuz, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PL-1296

Differential Revision: https://phab.corp.pixielabs.ai/D3065

GitOrigin-RevId: d09410b388ea550d31ceeb0bdc266bf99a32ba29