Analyzer support to make sure aggregate functions cannot be nested
Summary: add rule and analyzer invocation to make sure that aggregate functions aren't nested. This isn't really necessary at the moment, but will become necessary if we allow more complicated aggregates in the future and just makes sure we don't run into problems when that happens. This was previously handled with the ToProto calls for AggregateExpressions in the ir node.

Test Plan: Added analyzer test to make sure that this fails.

Reviewers: nserrino, zasgar, michelle, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3080

GitOrigin-RevId: f788aae7f3d7ed0bd172ee602550dbafee60ac49