Make UDFInfo an argument to LogicalPlanner constructor
Summary:
Make UDFInfo a part of the LogicalPlanner constructor and make it so that logical planner inits a registry_info from the udfinfo. This makes it so we can create a Logical Planner with the registry info and use that registry info for the duration of the compiler state object's lifetime.

Also updated cgo boundary so that we pass UDFInfo (serialized) into the constructor of the planner.

Test Plan: Tests should all work as before.

Reviewers: nserrino, zasgar, michelle, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3114

GitOrigin-RevId: 215a3880afc69e89c4cb13634e5d4b21c4370c5d