Revert the changes to the RegistryInfo pointer in compiler state because I misjudged the lifetime of that object.
Summary: I misjudged the lifetime of the RegistryInfo object. It turns out that it should last longer than the compiler state object, something that will be necessary with upcoming changes to the LogicalPlanner object.

Test Plan: tests pass with changes, this is a nearly 1 to 1 reversion of teh changes in a prev commit.

Reviewers: zasgar, nserrino, michelle, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3111

GitOrigin-RevId: e2e0666d7da34d9eea44e3c358eb3f3164b7251e