Add export of UDF funcs for go and new go-bindata rule.
Summary:
added genrule and bindata rule that export UDF funcs from the entirety of the system and puts them into a form that's accessible to go files.

New bindata comes from rules_go, but I wanted to add new features and the rule wasn't maintained to it so I co-opted it for our purposes.

Test Plan: .

Reviewers: #engineering, zasgar, michelle, nserrino

Reviewed By: #engineering, zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D3136

GitOrigin-RevId: b2010f1e40202bafccf3d250213eeaa58b1f80c3