[PL-1153] Make Limit node blocking
Summary: This fixes PL-1153  by making the limit node blocking. This way limit only returns the argument number of rows. I thought this was gonna be more complicated than it is, but this seems to be a really simple solution.

Test Plan: add test in splitter to make sure limit is blocking. Also tested in the UI.

Reviewers: zasgar, michelle, nserrino, #engineering

Reviewed By: zasgar, nserrino, #engineering

JIRA Issues: PL-1153

Differential Revision: https://phab.corp.pixielabs.ai/D3138

GitOrigin-RevId: 5068fe209223bde16d6be4a6737227d12230ee53